Rename UARTLib to PS5Lib, because it doesn't only do UART stuff.
Someone please give me a better name for this.
