+++ PS5Lib/NORData.cs
@@ -1,5 +1,5 @@
 using System.Text;
-using UARTLib;
+using PS5Lib;
 
 namespace PS5NORModifier;
