+++ PS5NORModifier.sln
@@ -7,7 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PS5NORModifier", "PS5NORMod
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UART-CL_By_TheCod3r", "UART-CL_By_TheCod3r\UART-CL_By_TheCod3r.csproj", "{0438F03C-BC9A-4F8E-88F3-CDAF252487C2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UARTLib", "UARTLib\UARTLib.csproj", "{AF657D71-B71E-4066-A312-91E5C797361E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PS5Lib", "PS5Lib\PS5Lib.csproj", "{AF657D71-B71E-4066-A312-91E5C797361E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
