+++ UART-CL_By_TheCod3r/UART-CL_By_TheCod3r.csproj
@@ -32,7 +32,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\UARTLib\UARTLib.csproj" />
+    <ProjectReference Include="..\PS5Lib\PS5Lib.csproj" />
   </ItemGroup>
 
 </Project>
